New configuration system (#17)
This commit rewrites the whole configuration system used in this project. Messages are now in English and everything (except for the potion recipe) is now fully configurable. Older configurations are going to be automatically migrated internally to the new config schema. Plugin will fail to load when a newer configuration file version is present than the current one.

* Rewrite config.yml file structure
* Translate default messages to English
* Fail to load when using newer config file
* Automatically migrate old configuration files

* Remove configuration description from `README`
* Update installation instructions in README